Move static instantiation method to AbstractViewController class
